% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m__nonzero_var}
\alias{w4m__nonzero_var}
\title{Support function to eliminate rows or columns that have zero variance}
\usage{
w4m__nonzero_var(m)
}
\arguments{
\item{m}{matrix: W4M data matrix potentially having rows or columns having zero variance}
}
\value{
matrix: input data matrix after removal of rows or columns having zero variance
}
\description{
(w4mclassfilter support function) Produce matrix from matrix xpre where all rows and columns having zero variance have been removed
}
\examples{

m <- matrix(
  c(
    1, 2, 3, 4,
    3, 3, 3, 4,
    5, 7, 11, 4,
    13, 17, 19, 4
  )
, nrow = 4
, ncol = 4
, byrow = TRUE
)
rownames(m) <- c("A", "B", "C", "D")
colnames(m) <- c("W", "X", "Y", "Z")
expected <- matrix(
  c(
    1, 2, 3,
    5, 7, 11,
    13, 17, 19
  )
  , nrow = 3
  , ncol = 3
  , byrow = TRUE
)
rownames(expected) <- c("A", "C", "D")
colnames(expected) <- c("W", "X", "Y")
all.equal(w4m__nonzero_var(m), expected)

}
