% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m_filter_by_sample_class}
\alias{w4m_filter_by_sample_class}
\title{Filter W4M Samples by Class of Sample}
\usage{
w4m_filter_by_sample_class(dataMatrix_in, sampleMetadata_in,
  variableMetadata_in, dataMatrix_out, sampleMetadata_out, variableMetadata_out,
  classes = c(), include = FALSE, class_column = "class",
  samplename_column = "sampleMetadata",
  data_imputation = w4m_filter_imputation)
}
\arguments{
\item{dataMatrix_in}{character: path to input file containing data matrix (tsv, rows are feature names, columns are sample names}

\item{sampleMetadata_in}{character: path to input file containing sample metadata (tsv, rows are sample names, one column's name matches class_column)}

\item{variableMetadata_in}{character: path to input file containing variable metadata (tsv, rows are variable names)}

\item{dataMatrix_out}{character: path to output file containing data matrix (tsv, rows are feature names, columns are sample names}

\item{sampleMetadata_out}{character: path to output file containing sample metadata (tsv, rows are sample names, one column's name matches class_column)}

\item{variableMetadata_out}{character: path to input file containing variable metadata (tsv, rows are variable names)}

\item{classes}{character array: names of sample classes to include or exclude; default is an empty array}

\item{include}{logical: TRUE, include named sample classes; FALSE (the default), exclude named sample classes}

\item{class_column}{character: name of "class" column, defaults to "class"}

\item{samplename_column}{character: name of column with sample name, defaults to "sampleMetadata"}

\item{data_imputation}{function(m): default imputation method for 'intb' data, where intensities have background subtracted - impute zero for NA}
}
\value{
logical: TRUE only if filtration succeeded
}
\description{
Filter set of retention-corrected W4M files (dataMatrix, sampleMetadata) by feature
}
\author{
Art Eschenlauer, \email{esch0041@umn.edu}
}
\concept{
w4m workflow4metabolomics
}
\keyword{multivariate}
