% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m_filter_median_imputation}
\alias{w4m_filter_median_imputation}
\title{Impute missing intensities to median for W4M data matrix}
\usage{
w4m_filter_median_imputation(m)
}
\arguments{
\item{m}{matrix: W4M data matrix potentially containing NA or negative values}
}
\value{
matrix: input data matrix with zeros substituted for negative values and median substituted for missing values
}
\description{
Substitute median feature intensity (across all samples) for missing values and zero for negative values in W4M data matrix
}
\examples{
# input contains negative and missing values
my_input <- matrix(c(NA,-1,3,2), ncol = 2, nrow = 2)

# expected output converts negative and missing values to zero
my_expected <- matrix(c(3,0,3,2), ncol = 2, nrow = 2)

# run the imputation method to generate actual output
my_output <- w4m_filter_median_imputation(my_input)

# validate actual output against expected output
all.equal(my_output, my_expected, check.attributes = FALSE)

}
\seealso{
\url{https://github.com/HegemanLab/w4mclassfilter}

\url{http://workflow4metabolomics.org/}
}
\author{
Art Eschenlauer, \email{esch0041@umn.edu}
}
\concept{w4m workflow4metabolomics}
