% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m__nonzero_var}
\alias{w4m__nonzero_var}
\title{Support function to eliminate rows or columns that have zero variance}
\usage{
w4m__nonzero_var(m)
}
\arguments{
\item{m}{matrix: W4M data matrix potentially having rows or columns having zero variance}
}
\value{
matrix: input data matrix after removal of rows or columns having zero variance
}
\description{
(w4mclassfilter support function) Produce matrix from matrix xpre where all rows and columns having zero variance have been removed
}
