% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m__var_by_rank_or_file}
\alias{w4m__var_by_rank_or_file}
\title{Support function to compute variances of matrix rows or columns}
\usage{
w4m__var_by_rank_or_file(m, dim = 1)
}
\arguments{
\item{m}{matrix: W4M data matrix for which variance must be computed for rows or columns}

\item{dim}{integer: For variances of rows, dim == 1, for variances of columns, dim == 2}
}
\value{
vector of numeric: variances for rows or columns
}
\description{
(w4mclassfilter support function) Compute variances of rows or columns of a W4M data matrix
}
\examples{

m <- base::matrix(
  c(
    1, 2, 3,
    5, 7, 11,
    13, 17, 19
  )
, nrow = 3
, ncol = 3
, byrow = TRUE
)
rowvars <- w4m__var_by_rank_or_file(m = m, dim = 1)
expecteds <- c(stats::var(c(1,2,3)),stats::var(c(5,7,11)),stats::var(c(13,17,19)))
base::all.equal(rowvars, expecteds)
colvars <- w4m__var_by_rank_or_file(m = m, dim = 2)
expecteds <- c(base::var(c(1,5,13)),base::var(c(2,7,17)),base::var(c(3,11,19)))
base::all.equal(colvars, expecteds)

}
\seealso{
\url{https://github.com/HegemanLab/w4mclassfilter}

\url{http://stackoverflow.com/a/25100036}
}
\author{
Art Eschenlauer, \email{esch0041@umn.edu}
}
\concept{
w4m workflow4metabolomics
}
