% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m_filter_by_sample_class}
\alias{w4m_filter_by_sample_class}
\title{Filter W4M Samples by Class of Sample}
\usage{
w4m_filter_by_sample_class(dataMatrix_in, sampleMetadata_in,
  variableMetadata_in, dataMatrix_out, sampleMetadata_out, variableMetadata_out,
  classes = c(), include = FALSE, class_column = "class",
  samplename_column = "sampleMetadata",
  data_imputation = w4m_filter_imputation, failure_action = print)
}
\arguments{
\item{dataMatrix_in}{character: path to input file containing data matrix (tsv, rows are feature names, columns are sample names}

\item{sampleMetadata_in}{character: path to input file containing sample metadata (tsv, rows are sample names, one column's name matches class_column)}

\item{variableMetadata_in}{character: path to input file containing variable metadata (tsv, rows are variable names)}

\item{dataMatrix_out}{character: path to output file containing data matrix (tsv, rows are feature names, columns are sample names}

\item{sampleMetadata_out}{character: path to output file containing sample metadata (tsv, rows are sample names, one column's name matches class_column)}

\item{variableMetadata_out}{character: path to input file containing variable metadata (tsv, rows are variable names)}

\item{classes}{character array: names of sample classes to include or exclude; default is an empty array}

\item{include}{logical: TRUE, include named sample classes; FALSE (the default), exclude named sample classes}

\item{class_column}{character: name of "class" column, defaults to "class"}

\item{samplename_column}{character: name of column with sample name, defaults to "sampleMetadata"}

\item{data_imputation}{function(m): default imputation method for 'intb' data, where intensities have background subtracted - impute zero for NA}

\item{failure_action}{function(x, ...): action to take upon failure - defaults to 'print(x,...)'}
}
\value{
logical: TRUE only if filtration succeeded
}
\description{
Filter set of retention-corrected W4M files (dataMatrix, sampleMetadata) by feature
}
\examples{
print(' # set the paths to your input files                                  ')
print(' dataMatrix_in <- "tests/testthat/input_dataMatrix.tsv"               ')
print(' sampleMetadata_in <- "tests/testthat/input_sampleMetadata.tsv"       ')
print(' variableMetadata_in <- "tests/testthat/input_variableMetadata.tsv"   ')
print('                                                                      ')
print(' # set the paths to your (nonexistent) output files                   ')
print(' #    in a directory that DOES need to exist                          ')
print(' dataMatrix_out <- "tests/testthat/output_dataMatrix.tsv"             ')
print(' sampleMetadata_out <- "tests/testthat/output_sampleMetadata.tsv"     ')
print(' variableMetadata_out <- "tests/testthat/output_variableMetadata.tsv" ')
print('                                                                      ')
print(' # Example: running the filter to exclude only unwanted samples       ')
print(' #   include = FALSE means exclude samples with class blankpos        ')
print(' w4m_filter_by_sample_class(                                          ')
print('   dataMatrix_in = dataMatrix_in                                      ')
print(' , dataMatrix_out = dataMatrix_out                                    ')
print(' , variableMetadata_in = variableMetadata_in                          ')
print(' , variableMetadata_out = variableMetadata_out                        ')
print(' , sampleMetadata_out = sampleMetadata_out                            ')
print(' , sampleMetadata_in = sampleMetadata_in                              ')
print(' , classes = c("blankpos")                                            ')
print(' , include = FALSE                                                    ')

}
\seealso{
\url{https://github.com/HegemanLab/w4mclassfilter}

\url{http://workflow4metabolomics.org/}
}
\author{
Art Eschenlauer, \email{esch0041@umn.edu}
}
\concept{
w4m workflow4metabolomics
}
\keyword{multivariate}
