% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassFilter.R
\name{w4m_filter_imputation}
\alias{w4m_filter_imputation}
\title{Impute W4M Data Matrix Values}
\usage{
w4m_filter_imputation(m)
}
\arguments{
\item{m}{matrix: W4M data matrix potentially containing NA or negative values}
}
\value{
matrix: input data matrix with zeros substituted for negative or NA values
}
\description{
Impute missing or negative values in W4M files as zero
}
\examples{
my_input <- matrix(c(NA,1,-1,2), ncol = 2, nrow = 2)
my_expected <- matrix(c(0,1,0,2), ncol = 2, nrow = 2)
my_output <- w4m_filter_imputation(my_input)
all.equal(my_output, my_expected, check.attributes = FALSE)

}
\seealso{
\url{https://github.com/HegemanLab/w4mclassfilter}

\url{http://workflow4metabolomics.org/}
}
\author{
Art Eschenlauer, \email{esch0041@umn.edu}
}
\concept{
w4m workflow4metabolomics
}
